Decompile imul, and follow the deferrals around it
Started with imul, which was unhandled, and then went through the TODOs
already standing in the code. Each one was covering a wrong answer rather
than a missing nicety.

imul writes a low and a high half; only the low one is ever used here.

An integer vector immediate was built with a float constructor, so
`& float2(0.000000, 0.000000)` was the mask 255 read as float bits.

lerp took its arguments in bytecode order, which puts the blend factor
first. HLSL wants lerp(from, to, amount), so every decompiled lerp was
blending the wrong pair.

`// TODO: emit an explicit cast rather than relying on implicit conversion`
covered ftoi, ftou, itof and utof, all four compiled as a plain move.
uint_convert was returning float4(v.y + v.x, ...) where the source casts
each to an integer first - the truncation simply gone, and compiling, which
is why nothing said so. Where the result has to be an integer it does not
compile at all: `10 * v & 7` is X3082. A cast binds tighter than the
arithmetic around it, so anything that is not a single term needs brackets,
and it is as wide as what it converts, `(float)` over two components being
X3014.

`// TODO: consider offset` was handled in one place out of three. A shader
sampling four ways - sample_l, gather4, ld and sample_c_lz, each with a
texel offset - lost every one of them, reading the wrong texels and
compiling while it did. AsmWriter matched on Sample alone, so the suffix
belongs on the mnemonic. The AST writer had three of the four and dropped
Load. `tex.Load(1)` stood behind that: the address is an int3, and
l(1, 2, 0, 0) came back as 1 because an immediate on an integer opcode was
built one component wide whatever the operand held - nor is it as wide as
the destination mask, since ld writes four components and addresses three.

`// TODO: mode` on dcl_sampler wrote mode_default whatever the sampler was,
and two fixtures declare comparison samplers.

`// TODO: determine vector size` sat over a CastOperation built as half4
whatever it was casting, so a two component partial precision write gave
float4(half4(saturate(texcoord.xy)), texcoord.z, 1) - X3014 twice over.
CastOperation was the same idea as ConvertOperation above and NodeCompiler
had no case for it at all, so it fell to the generic consumer branch and
came out as a constructor call.

The fixture added for that caught one more. `mad oC0.zw, v0.zz, c0.xy,
c0.yx` gave both addends as float2(1, 0). Reading a defined constant took
the first N entries of the swizzle, N being how many components are
written - but which entries are read depends on which components, not how
many, and a zw destination reads entries 2 and 3. Three fixtures were
wrong: an mad adding 1 to z instead of 0, and a mov of c0.yzw reading
(1, 2, 3) where c0 is (1, 2, 3, 4). All of them compiled, and the AST
writer had them right throughout - nothing compares the two writers, which
is how a wrong constant sat in a golden file.

+++ HlslDecompiler.Tests/ShaderSources/gs_4_1/circle.fx
@@ -19,7 +19,7 @@ void main(point GS_IN i[1], inout TriangleStream<GS_OUT> stream)
 	o.color = i[0].color;
 	stream.Append(o);
 	for (int t0 = 1; t0 <= 17; t0 = t0 + 1) {
-		o.sv_position = float4(0.5 * cos(0.392699093 * t0) + i[0].sv_position.x, 0.5 * sin(0.392699093 * t0) + i[0].sv_position.y, i[0].sv_position.zw);
+		o.sv_position = float4(0.5 * cos(0.392699093 * (float)t0) + i[0].sv_position.x, 0.5 * sin(0.392699093 * (float)t0) + i[0].sv_position.y, i[0].sv_position.zw);
 		o.color = i[0].color;
 		stream.Append(o);
 	}
