+++ HlslDecompiler.Tests/ShaderSources/ps_4_0/immediate_constant_buffer.fx
@@ -15,7 +15,7 @@ float4 main(float2 texcoord : TEXCOORD) : SV_Target
 {
 	float4 t0 = 0;
 	for (int t1 = 0; t1 < 4; t1 = t1 + 1) {
-		t0 = layers.Sample(samp, float3(texcoord, t1)) * dot(weights, icb[t1]) + t0;
+		t0 = layers.Sample(samp, float3(texcoord, (float)t1)) * dot(weights, icb[t1]) + t0;
 	}
 	return t0;
 }
