+++ HlslDecompiler.Tests/ShaderSources/ps_4_0/uint_convert.fx
@@ -2,5 +2,5 @@ float4 v;
 
 float4 main() : SV_Target
 {
-	return float4(v.y + v.x, v.x * 2, v.x, trunc(v.y));
+	return float4((float)((int)v.y + (uint)v.x), (float)((uint)v.x * 2), (float)(uint)v.x, trunc(v.y));
 }
